Create R Project-generate bar graphs-Microbiome
This was an R Project I made for summer of 2019 where I generated bar graphs for 5 metabolites. I alternated the shades of the bars so I could get wild type vs disease type. I ran a two-sided paired t-test in excel and just put significance bars on top of the bargraphs.